Start fixing structure, examples
