+++ examples/two_screens/main.rs
@@ -2,14 +2,15 @@
 mod examples_common;
 
 use bevy::{
+    ecs::query::ReadOnlyWorldQuery,
     prelude::*,
     render::camera::RenderTarget,
-    window::{WindowRef, WindowResolution},
+    window::{PrimaryWindow, WindowRef, WindowResolution},
 };
 
-use bevy_vfx_bag::post_processing2::v3::{
-    blur::Blur, chromatic_aberration::ChromaticAberration, flip::Flip, lut::LutSettings,
-    masks::Mask, pixelate::Pixelate, raindrops::Raindrops, PostProcessingPlugin, VfxOrdering,
+use bevy_vfx_bag::post_processing::{
+    blur::Blur, chromatic_aberration::ChromaticAberration, flip::Flip, lut::Lut, masks::Mask,
+    pixelate::Pixelate, raindrops::Raindrops, PostProcessingOrder, PostProcessingPlugin,
 };
 
 const NUM_EFFECTS: usize = 7;
@@ -18,38 +19,35 @@ const NUM_EFFECTS: usize = 7;
 struct Effects([&'static str; NUM_EFFECTS]);
 
 impl Effects {
-    fn insert_default(name: &str, priority: f32, entity: Entity, commands: &mut Commands) {
+    fn insert_default(name: &str, order: f32, entity: Entity, commands: &mut Commands) {
         if name == "Pixelate" {
             commands
                 .entity(entity)
-                .insert((Pixelate::default(), VfxOrdering::<Pixelate>::new(priority)));
+                .insert(Pixelate::default().with_order(order));
         } else if name == "Raindrops" {
-            commands.entity(entity).insert((
-                Raindrops::default(),
-                VfxOrdering::<Raindrops>::new(priority),
-            ));
+            commands
+                .entity(entity)
+                .insert(Raindrops::default().with_order(order));
         } else if name == "Flip" {
             commands
                 .entity(entity)
-                .insert((Flip::default(), VfxOrdering::<Flip>::new(priority)));
+                .insert(Flip::default().with_order(order));
         } else if name == "Mask" {
             commands
                 .entity(entity)
-                .insert((Mask::default(), VfxOrdering::<Mask>::new(priority)));
+                .insert(Mask::default().with_order(order));
         } else if name == "Lut" {
-            commands.entity(entity).insert((
-                LutSettings::default(),
-                VfxOrdering::<LutSettings>::new(priority),
-            ));
+            commands
+                .entity(entity)
+                .insert(Lut::default().with_order(order));
         } else if name == "Blur" {
             commands
                 .entity(entity)
-                .insert((Blur::default(), VfxOrdering::<Blur>::new(priority)));
+                .insert(Blur::default().with_order(order));
         } else if name == "ChromaticAberration" {
-            commands.entity(entity).insert((
-                ChromaticAberration::default(),
-                VfxOrdering::<ChromaticAberration>::new(priority),
-            ));
+            commands
+                .entity(entity)
+                .insert(ChromaticAberration::default().with_order(order));
         } else {
             panic!("Unknown effect name");
         }
@@ -65,7 +63,7 @@ impl Effects {
         } else if name == "Mask" {
             commands.entity(entity).remove::<Mask>();
         } else if name == "Lut" {
-            commands.entity(entity).remove::<LutSettings>();
+            commands.entity(entity).remove::<Lut>();
         } else if name == "Blur" {
             commands.entity(entity).remove::<Blur>();
         } else if name == "ChromaticAberration" {
@@ -94,8 +92,8 @@ fn main() {
         .add_startup_system(setup)
         .add_system(change_selection)
         .add_system(update_text)
-        .add_system(update_effects::<Window1>)
-        .add_system(update_effects::<Window2>);
+        .add_system(update_effects::<With<PrimaryWindow>>)
+        .add_system(update_effects::<Without<PrimaryWindow>>);
     // .add_system(update_effects_order);
 
     // let s = bevy_mod_debugdump::get_render_schedule(&mut app);
@@ -109,34 +107,49 @@ fn main() {
     app.run();
 }
 
-#[derive(Component, Clone, Copy)]
-struct Window1;
-
-#[derive(Component, Clone, Copy)]
-struct Window2;
-
 fn setup(mut commands: Commands, asset_server: Res<AssetServer>, effects: Res<Effects>) {
     let vfx_bundle = (
         Pixelate::default(),
         Raindrops::default(),
         Flip::default(),
         Mask::default(),
-        LutSettings::default(),
+        Lut::default(),
         Blur::default(),
         ChromaticAberration::default(),
     );
 
-    commands.spawn((
-        Camera3dBundle {
-            transform: Transform::from_xyz(-5.0, 12., 10.0)
-                .looking_at(Vec3::new(0., 1., 0.), Vec3::Y),
-            ..default()
-        },
-        vfx_bundle.clone(),
-        Window1,
-    ));
+    commands
+        .spawn((
+            Camera3dBundle {
+                transform: Transform::from_xyz(-5.0, 12., 10.0)
+                    .looking_at(Vec3::new(0., 1., 0.), Vec3::Y),
+                ..default()
+            },
+            vfx_bundle.clone(),
+            NodeBundle {
+                style: Style {
+                    size: Size::new(Val::Percent(100.0), Val::Percent(100.0)),
+                    align_items: AlignItems::FlexStart,
+                    padding: UiRect::all(Val::Px(10.0)),
+                    justify_content: JustifyContent::FlexStart,
+                    flex_direction: FlexDirection::Column,
+                    ..default()
+                },
+                ..default()
+            },
+        ))
+        .with_children(|parent| {
+            let list = examples_common::List::new(
+                asset_server.load("fonts/FiraSans-Bold.ttf"),
+                effects.0.iter(),
+            );
+            let text_bundle = list.as_text_bundle();
 
-    let window_id = commands
+            parent.spawn(text_bundle).insert(list);
+        });
+
+    // let primary_window = WindowRef::default();
+    let secondary_window = commands
         .spawn(Window {
             resolution: WindowResolution::new(0.4, 0.4),
             ..default()
@@ -144,70 +157,61 @@ fn setup(mut commands: Commands, asset_server: Res<AssetServer>, effects: Res<Ef
         .id();
 
     // second window camera
-    commands.spawn((
-        Camera3dBundle {
-            transform: Transform::from_xyz(-5.0, 12., 10.0)
-                .looking_at(Vec3::new(0., 1., 0.), Vec3::Y),
-            camera: Camera {
-                target: RenderTarget::Window(WindowRef::Entity(window_id)),
-                ..default()
-            },
-            ..default()
-        },
-        vfx_bundle,
-        Window2,
-    ));
-
     commands
-        .spawn((NodeBundle {
-            style: Style {
-                size: Size::new(Val::Percent(100.0), Val::Percent(100.0)),
-                align_items: AlignItems::FlexStart,
-                padding: UiRect::all(Val::Px(10.0)),
-                justify_content: JustifyContent::FlexStart,
-                flex_direction: FlexDirection::Column,
+        .spawn((
+            Camera3dBundle {
+                transform: Transform::from_xyz(-5.0, 12., 10.0)
+                    .looking_at(Vec3::new(0., 1., 0.), Vec3::Y),
+                camera: Camera {
+                    target: RenderTarget::Window(WindowRef::Entity(secondary_window)),
+                    ..default()
+                },
                 ..default()
             },
-            ..default()
-        },))
+            vfx_bundle,
+            NodeBundle {
+                style: Style {
+                    size: Size::new(Val::Percent(100.0), Val::Percent(100.0)),
+                    align_items: AlignItems::FlexStart,
+                    padding: UiRect::all(Val::Px(10.0)),
+                    justify_content: JustifyContent::FlexStart,
+                    flex_direction: FlexDirection::Column,
+                    ..default()
+                },
+                ..default()
+            },
+        ))
         .with_children(|parent| {
-            let vfx = examples_common::List::new(
+            let list = examples_common::List::new(
                 asset_server.load("fonts/FiraSans-Bold.ttf"),
                 effects.0.iter(),
             );
-            let tb = vfx.as_text_bundle();
+            let text_bundle = list.as_text_bundle();
 
-            parent
-                .spawn(tb.clone())
-                .insert((vfx.clone(), Window1, WindowRelation(window_id)));
-            parent
-                .spawn(tb)
-                .insert((vfx, Window2, WindowRelation(window_id)));
+            parent.spawn(text_bundle).insert(list);
         });
 }
 
-#[derive(Component, Deref)]
-struct WindowRelation(Entity);
-
 fn change_selection(
     keyboard_input: Res<Input<KeyCode>>,
-    mut query: Query<(&WindowRelation, &mut examples_common::List)>,
+    mut query: Query<&mut examples_common::List>,
     windows: Query<(Entity, &Window)>,
 ) {
     for (entity, window) in windows.iter() {
         if window.focused {
-            for (window_id, mut list) in query.iter_mut() {
-                if **window_id == entity {
-                    if keyboard_input.just_pressed(KeyCode::Space) {
-                        list.toggle_selected();
-                    } else if keyboard_input.just_pressed(KeyCode::Up) {
-                        list.up();
-                    } else if keyboard_input.just_pressed(KeyCode::Down) {
-                        list.down();
-                    } else if keyboard_input.just_pressed(KeyCode::T) {
-                        list.toggle_enabled();
-                    }
-                }
+            for (mut list) in query.iter_mut() {
+                todo!()
+                // if **window_id == entity {
+                //     if keyboard_input.just_pressed(KeyCode::Space) {
+                //         list.toggle_selected();
+                //     } else if keyboard_input.just_pressed(KeyCode::Up) {
+                //         list.up();
+                //     } else if keyboard_input.just_pressed(KeyCode::Down) {
+                //         list.down();
+                //     } else if keyboard_input.just_pressed(KeyCode::T) {
+                //         list.toggle_enabled();
+                //     }
+                // }
             }
         }
     }
@@ -219,10 +223,10 @@ fn update_text(mut query: Query<(&mut Text, &examples_common::List)>) {
     }
 }
 
-fn update_effects<W: Component>(
+fn update_effects<F: ReadOnlyWorldQuery>(
     mut commands: Commands,
-    window_list: Query<&examples_common::List, (With<W>, Changed<examples_common::List>)>,
-    window_camera: Query<Entity, (With<W>, With<Camera>)>,
+    window_list: Query<&examples_common::List, (F, Changed<examples_common::List>)>,
+    window_camera: Query<Entity, (F, With<Camera>)>,
 ) {
     let mut priority = 0.0;
