+++ src/post_processing/pixelate.rs
@@ -17,66 +17,38 @@ pub(crate) use bevy::{
     },
 };
 
-use crate::post_processing2::v3::{DrawPostProcessing, UniformBindGroup};
+use crate::post_processing::{DrawPostProcessingEffect, UniformBindGroup};
 
-use super::{DrawPostProcessingEffect, PostProcessingPhaseItem, VfxOrdering};
+use super::{PostProcessingPhaseItem, VfxOrdering};
 
-const WAVE_SHADER_HANDLE: HandleUntyped =
-    HandleUntyped::weak_from_u64(Shader::TYPE_UUID, 1792660281364049744);
-
-/// Wave parameters.
-///
-/// Note that the parameters for the X axis causes a wave effect
-/// towards the left- and right sides of the screen.
-/// For example, if we have 1 wave in the X axis,
-/// we will have one part of the screen stretched towards the right
-/// horizontally, and one part stretched towards the left.
-#[derive(Default, Debug, Copy, Clone, Component, ShaderType)]
-pub struct Wave {
-    /// How many waves in the x axis.
-    pub waves_x: f32,
-
-    /// How many waves in the y axis.
-    pub waves_y: f32,
-
-    /// How fast the x axis waves oscillate.
-    pub speed_x: f32,
-
-    /// How fast the y axis waves oscillate.
-    pub speed_y: f32,
-
-    /// How much displacement the x axis waves cause.
-    pub amplitude_x: f32,
-
-    /// How much displacement the y axis waves cause.
-    pub amplitude_y: f32,
-}
+pub(crate) const PIXELATE_SHADER_HANDLE: HandleUntyped =
+    HandleUntyped::weak_from_u64(Shader::TYPE_UUID, 11093977931118718560);
 
 #[derive(Resource)]
-pub(crate) struct WaveData {
+pub(crate) struct PixelateData {
     pub pipeline_id: CachedRenderPipelineId,
     pub uniform_layout: BindGroupLayout,
 }
 
-impl FromWorld for WaveData {
+impl FromWorld for PixelateData {
     fn from_world(world: &mut World) -> Self {
         let (uniform_layout, pipeline_id) = super::create_layout_and_pipeline(
             world,
-            "Wave",
+            "Pixelate",
             &[BindGroupLayoutEntry {
                 binding: 0,
                 ty: BindingType::Buffer {
                     ty: BufferBindingType::Uniform,
                     has_dynamic_offset: true,
-                    min_binding_size: Some(Wave::min_size()),
+                    min_binding_size: Some(Pixelate::min_size()),
                 },
                 visibility: ShaderStages::FRAGMENT,
                 count: None,
             }],
-            WAVE_SHADER_HANDLE.typed(),
+            PIXELATE_SHADER_HANDLE.typed(),
         );
 
-        WaveData {
+        PixelateData {
             pipeline_id,
             uniform_layout,
         }
@@ -88,39 +60,45 @@ impl bevy::prelude::Plugin for Plugin {
     fn build(&self, app: &mut App) {
         load_internal_asset!(
             app,
-            WAVE_SHADER_HANDLE,
-            concat!(env!("CARGO_MANIFEST_DIR"), "/assets/shaders/", "wave.wgsl"),
+            PIXELATE_SHADER_HANDLE,
+            concat!(
+                env!("CARGO_MANIFEST_DIR"),
+                "/assets/shaders/",
+                "pixelate.wgsl"
+            ),
             Shader::from_wgsl
         );
 
         // This puts the uniform into the render world.
-        app.add_plugin(ExtractComponentPlugin::<Wave>::default())
-            .add_plugin(UniformComponentPlugin::<Wave>::default());
+        app.add_plugin(ExtractComponentPlugin::<Pixelate>::default())
+            .add_plugin(UniformComponentPlugin::<Pixelate>::default());
 
         super::render_app(app)
             .add_system_to_stage(
                 RenderStage::Extract,
-                super::extract_post_processing_camera_phases::<Wave>,
+                super::extract_post_processing_camera_phases::<Pixelate>,
             )
-            .init_resource::<WaveData>()
-            .init_resource::<UniformBindGroup<Wave>>()
+            .init_resource::<PixelateData>()
+            .init_resource::<UniformBindGroup<Pixelate>>()
             .add_system_to_stage(RenderStage::Prepare, prepare)
             .add_system_to_stage(RenderStage::Queue, queue)
-            .add_render_command::<PostProcessingPhaseItem, DrawPostProcessingEffect<Wave>>();
+            .add_render_command::<PostProcessingPhaseItem, DrawPostProcessingEffect<Pixelate>>();
     }
 }
 
 fn prepare(
-    data: Res<WaveData>,
+    data: Res<PixelateData>,
     mut views: Query<(
         Entity,
         &mut RenderPhase<PostProcessingPhaseItem>,
-        &VfxOrdering<Wave>,
+        &VfxOrdering<Pixelate>,
     )>,
     draw_functions: Res<DrawFunctions<PostProcessingPhaseItem>>,
 ) {
     for (entity, mut phase, order) in views.iter_mut() {
-        let draw_function = draw_functions.read().id::<DrawPostProcessingEffect<Wave>>();
+        let draw_function = draw_functions
+            .read()
+            .id::<DrawPostProcessingEffect<Pixelate>>();
 
         phase.add(PostProcessingPhaseItem {
             entity,
@@ -133,17 +111,17 @@ fn prepare(
 
 fn queue(
     render_device: Res<RenderDevice>,
-    data: Res<WaveData>,
-    mut bind_group: ResMut<UniformBindGroup<Wave>>,
-    uniforms: Res<ComponentUniforms<Wave>>,
-    views: Query<Entity, With<Wave>>,
+    data: Res<PixelateData>,
+    mut bind_group: ResMut<UniformBindGroup<Pixelate>>,
+    uniforms: Res<ComponentUniforms<Pixelate>>,
+    views: Query<Entity, With<Pixelate>>,
 ) {
     bind_group.inner = None;
 
     if let Some(uniforms) = uniforms.binding() {
         if !views.is_empty() {
             bind_group.inner = Some(render_device.create_bind_group(&BindGroupDescriptor {
-                label: Some("Wave Uniform Bind Group"),
+                label: Some("Pixelate Uniform Bind Group"),
                 layout: &data.uniform_layout,
                 entries: &[BindGroupEntry {
                     binding: 0,
@@ -154,7 +132,20 @@ fn queue(
     }
 }
 
-impl ExtractComponent for Wave {
+/// TODO
+#[derive(Debug, ShaderType, Component, Clone, Copy)]
+pub struct Pixelate {
+    /// TODO
+    pub block_size: f32,
+}
+
+impl Default for Pixelate {
+    fn default() -> Self {
+        Self { block_size: 8.0 }
+    }
+}
+
+impl ExtractComponent for Pixelate {
     type Query = (&'static Self, &'static Camera);
     type Filter = ();
     type Out = Self;
