--- src/post_processing2.rs
@@ -1,5 +1,0 @@
-/// Helpers for implementing post processing effects.
-pub(crate) mod util;
-
-/// WE GO AGAN
-pub mod v3;
