+++ src/util.rs
@@ -157,3 +157,25 @@ macro_rules! load_lut {
         (handle, IsFixed(false))
     }};
 }
+
+/// Macro for selecting a way to load shaders
+/// based on the "dev" feature.
+/// If the feature is on, we load from an assets-relative
+/// path. Suitable for hot-reloading.
+/// Else, the shader is loaded via the handle.
+/// Suitable when this crate is used as a dependency.
+#[doc(hidden)]
+#[macro_export]
+macro_rules! shader_ref {
+    ($handle: ident, $path_str: expr) => {{
+        use bevy::render::render_resource::ShaderRef;
+
+        let s: ShaderRef = if cfg!(feature = "dev") {
+            $path_str.into()
+        } else {
+            $handle.typed().into()
+        };
+
+        s
+    }};
+}
